digital-credentials/digital-credential-client-allows-protocol.https.html

<!DOCTYPE html>
<title>Digital Credential Client Allows Protocol method.</title>
<link rel="help" href="https://w3c-fedid.github.io/digital-credentials/" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body></body>
<script type="module">

  promise_test(async (t) => {
    const result = DigitalCredential.clientAllowsProtocol('protocol.banana');
    assert_equals(result, false);
  }, "Client does not allow protocol 'protocol.banana'.");

</script>
